fix Modal passive wheel event & Modal proptypes & directory structure
